# packages/sui-media-api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/media": {
      "get": {
        "operationId": "MediaController_findAll",
        "summary": "List media",
        "description": "List media with filtering, search, sorting, and pagination. Supports both offset and cursor-based pagination.",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 0,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of items to skip",
            "schema": {
              "minimum": 0,
              "default": 0,
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for cursor-based pagination",
            "schema": {
              "example": "eyJpZCI6Im1lZGlhXzEyMyJ9",
              "type": "string"
            }
          },
          {
            "name": "mediaType",
            "required": false,
            "in": "query",
            "description": "Filter by media type",
            "schema": {
              "example": "video",
              "enum": [
                "image",
                "video",
                "album"
              ],
              "type": "string"
            }
          },
          {
            "name": "mime",
            "required": false,
            "in": "query",
            "description": "Filter by specific MIME type",
            "schema": {
              "example": "video/mp4",
              "type": "string"
            }
          },
          {
            "name": "uploadedBy",
            "required": false,
            "in": "query",
            "description": "Filter by author user ID",
            "schema": {
              "example": "user_123",
              "type": "string"
            }
          },
          {
            "name": "publicity",
            "required": false,
            "in": "query",
            "description": "Filter by publicity level",
            "schema": {
              "example": "public",
              "enum": [
                "public",
                "private",
                "paid",
                "subscription"
              ],
              "type": "string"
            }
          },
          {
            "name": "rating",
            "required": false,
            "in": "query",
            "description": "Filter by content rating",
            "schema": {
              "example": "ga",
              "enum": [
                "ga",
                "nc17"
              ],
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Filter by date range start (ISO date)",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Filter by date range end (ISO date)",
            "schema": {
              "example": "2024-12-31T23:59:59.999Z",
              "type": "string"
            }
          },
          {
            "name": "mediaClass",
            "required": false,
            "in": "query",
            "description": "Filter by MediaClass ID",
            "schema": {
              "example": "class_123",
              "type": "string"
            }
          },
          {
            "name": "uploadSessionId",
            "required": false,
            "in": "query",
            "description": "Filter by upload session",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in filename, title, description, tags",
            "schema": {
              "example": "sunset beach",
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "Comma-separated tag search",
            "schema": {
              "example": "nature,ocean,sunset",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sort field (prefix with - for descending)",
            "schema": {
              "default": "-createdAt",
              "example": "-createdAt",
              "enum": [
                "createdAt",
                "updatedAt",
                "size",
                "title",
                "views",
                "-createdAt",
                "-updatedAt",
                "-size",
                "-title",
                "-views"
              ],
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft-deleted items",
            "schema": {
              "default": false,
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMediaResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required"
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "post": {
        "operationId": "MediaController_create",
        "summary": "Create media entry",
        "description": "Create a new media entry in the database. Used after file upload to create the metadata record.",
        "parameters": [
          {
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Media creation data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMediaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Media created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "401": {
            "description": "Unauthorized - authentication required"
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/media/{id}": {
      "get": {
        "operationId": "MediaController_findOne",
        "summary": "Get media by ID",
        "description": "Get a single media item by ID. Increments view count on each access.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Media ID",
            "schema": {
              "example": "media_1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media item retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "404": {
            "description": "Media not found"
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "MediaController_update",
        "summary": "Update media",
        "description": "Update media metadata. Only the owner can update their media.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Media ID",
            "schema": {
              "example": "media_1",
              "type": "string"
            }
          },
          {
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Fields to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Media updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "403": {
            "description": "Forbidden - user is not the owner"
          },
          "404": {
            "description": "Media not found"
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "MediaController_remove",
        "summary": "Delete media",
        "description": "Delete media (file + database). Default is soft delete. Use hardDelete=true query parameter for permanent deletion.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Media ID",
            "schema": {
              "example": "media_1",
              "type": "string"
            }
          },
          {
            "name": "hardDelete",
            "required": false,
            "in": "query",
            "description": "Whether to permanently delete (default: false)",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Media deleted successfully"
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "403": {
            "description": "Forbidden - user is not the owner"
          },
          "404": {
            "description": "Media not found"
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/media/{id}/restore": {
      "post": {
        "operationId": "MediaController_restore",
        "summary": "Restore deleted media",
        "description": "Restore a soft-deleted media item. Only the owner can restore their media.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Media ID",
            "schema": {
              "example": "media_1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media restored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication required"
          },
          "403": {
            "description": "Forbidden - user is not the owner"
          },
          "404": {
            "description": "Media not found"
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/media/{id}/extract-metadata": {
      "post": {
        "operationId": "MediaController_extractMetadata",
        "summary": "Extract media metadata",
        "description": "Extract metadata from a media file (video/audio/image properties).",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Media ID",
            "schema": {
              "example": "media_1",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metadata extracted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractMetadataResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Unable to extract metadata"
          },
          "404": {
            "description": "Media not found"
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
    "/media/{id}/generate-thumbnail": {
      "post": {
        "operationId": "MediaController_generateThumbnail",
        "summary": "Generate video thumbnail",
        "description": "Generate a thumbnail for a video at a specific timestamp.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Media ID",
            "schema": {
              "example": "media_1",
              "type": "string"
            }
          },
          {
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Thumbnail generation parameters",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateThumbnailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thumbnail generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThumbnailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or media type"
          },
          "404": {
            "description": "Media not found"
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
    "/media/{id}/generate-sprites": {
      "post": {
        "operationId": "MediaController_generateSpriteSheet",
        "summary": "Generate video sprite sheet",
        "description": "Generate a sprite sheet for video scrubbing with VTT file.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Media ID",
            "schema": {
              "example": "media_1",
              "type": "string"
            }
          },
          {
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Sprite sheet generation parameters",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSpriteSheetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sprite sheet generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpriteSheetResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or media type"
          },
          "404": {
            "description": "Media not found"
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
...
  },
  "info": {
    "title": "Stoked UI Media API",
    "description": "REST API for Stoked UI Media Components - Upload, manage, and process media files with support for images, videos, and albums. Features include multipart uploads, metadata extraction, thumbnail generation, and comprehensive media management.",
    "version": "0.1.0",
    "contact": {
      "name": "Brian Stoker",
      "url": "https://github.com/stoked-ui/sui",
      "email": "brian@stoked-ui.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    },
    {
      "url": "https://api.stoked-ui.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "PaginatedMediaResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Array of media items",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of items",
            "example": 100
          },
          "limit": {
            "type": "number",
            "description": "Number of items per page",
            "example": 20
          },
          "offset": {
            "type": "number",
            "description": "Number of items skipped",
            "example": 0
          },
          "hasMore": {
            "type": "boolean",
            "description": "Whether there are more items",
            "example": true
          },
          "cursor": {
            "type": "string",
            "description": "Next cursor for cursor-based pagination",
            "example": "eyJpZCI6Im1lZGlhXzEyMyJ9"
          }
        },
        "required": [
          "data",
          "total",
          "limit",
          "offset",
          "hasMore"
        ]
      },
...
      "GenerateThumbnailDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "number",
            "description": "Timestamp in seconds where to capture the thumbnail",
            "example": 60.5,
            "minimum": 0
          },
          "size": {
            "type": "string",
            "description": "Thumbnail size preset",
            "enum": [
              "small",
              "medium",
              "large"
            ],
            "example": "medium",
            "default": "medium"
          }
        },
        "required": [
          "timestamp"
        ]
      },
...
    }
  }
}